requirements.txt
pyxdg
requests
python-dotenv
